+++ flake.nix
@@ -207,7 +207,7 @@
       system = "x86_64-linux";
       nixos = true;
       modules = builtins.attrNames {
-        inherit (modules) i3;
+        inherit (modules) i3 vncserver;
       };
     }
   ]) // (