+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `vncserver` module
 - Fixed rootless container ports not accessible through firewall
 - Use `podman-compose` instead of `docker-compose` for rootless containers
 - Added `contacts` module